split ProposalsListItem to seprate file
