+++ app/components/views/GovernancePage/Proposals/ProposalsList/ProposalsList.jsx
@@ -1,12 +1,5 @@
-import { FormattedMessage as T } from "react-intl";
-import { VotingProgress } from "indicators";
 import { PoliteiaLoading, NoProposals } from "indicators";
-import {
-  VOTESTATUS_ACTIVEVOTE,
-  VOTESTATUS_FINISHEDVOTE
-} from "actions/GovernanceActions";
 import InfiniteScroll from "react-infinite-scroller";
-import { FormattedRelative } from "shared";
 import { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { fetchMachine } from "stateMachines/FetchStateMachine";
@@ -14,79 +7,8 @@ import { useMachine } from "@xstate/react";
 import * as sel from "selectors";
 import * as gov from "actions/GovernanceActions";
 import { usePrevious } from "helpers";
+import ProposalsListItem from "../ProposalsListItem/ProposalsListItem";
 import styles from "./ProposalsList.module.css";
-
-function ProposalListItem({
-  name,
-  timestamp,
-  token,
-  voteCounts,
-  voteStatus,
-  currentVoteChoice,
-  quorumPass,
-  voteResult,
-  modifiedSinceLastAccess,
-  votingSinceLastAccess,
-  quorumMinimumVotes
-}) {
-  // TODO: add custom hook and move to separate file
-  const tsDate = useSelector((state) => sel.tsDate(state));
-  const dispatch = useDispatch();
-  const isVoting = voteStatus == VOTESTATUS_ACTIVEVOTE;
-  const modifiedClassName =
-    (!isVoting && modifiedSinceLastAccess) ||
-    (isVoting && votingSinceLastAccess)
-      ? "proposal-modified-since-last-access"
-      : null;
-
-  return (
-    <div
-      onClick={() => dispatch(gov.viewProposalDetails(token))}
-      className={[
-        "is-row",
-        "proposal-list-item",
-        voteResult,
-        modifiedClassName
-      ].join(" ")}>
-      <div className="info">
-        <div className="proposal-name">{name}</div>
-        <div className="proposal-token">{token}</div>
-      </div>
-      <div className="proposal-results-area">
-        {(voteStatus === VOTESTATUS_ACTIVEVOTE ||
-          voteStatus === VOTESTATUS_FINISHEDVOTE) && (
-          <div className="is-row voting-indicator">
-            <div
-              className={
-                "vote-choice " + (currentVoteChoice && currentVoteChoice.id)
-              }
-            />
-            <VotingProgress {...{ voteCounts, quorumMinimumVotes }} />
-          </div>
-        )}
-        {voteStatus !== VOTESTATUS_FINISHEDVOTE ? (
-          <div className="proposal-timestamp">
-            <T
-              id="proposalItem.lastUpdatedAt"
-              m="Last Updated {reldate}"
-              values={{
-                reldate: <FormattedRelative value={tsDate(timestamp)} />
-              }}
-            />
-          </div>
-        ) : (
-          <div className="vote-result">
-            {quorumPass ? (
-              voteResult
-            ) : (
-              <T id="proposals.quorumNotMet" m="Quorum not met" />
-            )}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
 
 // TODO: Get proposallistpagesize from politeia's request: /v1/policy
 async function onLoadMoreProposals(
@@ -182,7 +104,7 @@ export function ProposalList({ finishedVote, tab }) {
           threshold={300}>
           <div className={"proposal-list " + (finishedVote && "ended")}>
             {proposals[tab].map((v) => (
-              <ProposalListItem key={v.token} {...v} />
+              <ProposalsListItem key={v.token} {...v} />
             ))}
           </div>
         </InfiniteScroll>
