+++ app/components/views/GovernancePage/Proposals/ProposalsList/ProposalsList.module.css
@@ -1,0 +1,4 @@
+.loadingPage {
+  width: 100%;
+  text-align: center;
+}
