+++ app/components/views/GovernancePage/Proposals/ProposalsTab.module.css
@@ -1,8 +1,3 @@
-.loadingPage {
-  width: 100%;
-  text-align: center;
-}
-
 .politeiaButton {
   font-size: 13px !important;
   line-height: 17px !important;
